1.区分prod 生产环境。 2.优化maven依赖。 3.兼容releaseDao 1.1.2版本。 4.更改 缩略图接口。

+++ comm-util/src/main/java/com/wen/commutil/vo/PageVO.java
@@ -17,17 +17,17 @@ import java.util.List;
 public class PageVO<T> {
     private List<T> content;
 
-    private Long page;
+    private Integer page;
 
     private Integer size;
 
-    private Long total;
+    private Integer total;
 
-    public static <T> PageVO<T> of(List<T> content, Long page, Integer size, Long total) {
+    public static <T> PageVO<T> of(List<T> content, Integer page, Integer size, Integer total) {
         return new PageVO<>(content, page, size, total);
     }
 
-    public PageVO(List<T> content, Long page, Integer size, Long total) {
+    public PageVO(List<T> content, Integer page, Integer size, Integer total) {
         this.content = content;
         this.page = page;
         this.size = size;
